refactor: sort schedules by last execution by default
